removed all subscription and wait and pop related code
After we added multithread support some features (like diff,
subscription window and wait and pop) proved to be hard to implement in
a multithreaded tio. Now we have lots of deadlocks hapenning and we need
to fix the way subscriptions are handled to fix the deadlocks. This
commit is the first step: we removed subscription related code to
reimplement it in a thread safe way. We also removed not used features
(wait and pop, diff) so we can reimplement them in a thread safe way as
well.